feat: fix a little bug

+++ predict.py
@@ -157,7 +157,7 @@ for i in range(x_size):
 
 map_pic = integrate_tiles(in_path, tile_files)
 w, h, _ = map_pic.shape
-map_pic_normal_size = cv2.resize(map_pic, (w / 2, h / 2), interpolation=cv2.INTER_NEAREST)
+map_pic_normal_size = cv2.resize(map_pic, (int(w / 2), int(h / 2)), interpolation=cv2.INTER_NEAREST)
 cv2.imwrite(opt.RESULT_PATH, map_pic_normal_size)
 
 # for root, dirs, files in os.walk(web_dir, topdown=False):
